New DAO token mint mechanism
From now on, clan members will mint as much FDAO tokens as they received as clan reward. All clan members will split all the clan rewards. FDAO token supply will be equal to claimed clan member rewards.

With this new system, we no longer need to specify and create a proposal to give permission to mint new DAO tokens. We are getting rid of a heavy operational work.

And clans represent 50% of the DAO and the lords represent the remaining 50%